move time parsing to epoch subpkg
