+++ timetracker/epoch/calc.py
@@ -15,8 +15,38 @@ __author__ = "DV Klopfenstein, PhD"
 
 from datetime import datetime
 from datetime import timedelta
+from timetracker.consts import FMTDT
+from timetracker.consts import FMTDT24HMS
+STRPTIME = datetime.strptime
 
 
+# --------------------------------------------------------------
+def dt_from_str(txt):
+    """Get a datetime object, given a string"""
+    return STRPTIME(txt, FMTDT) if len(txt) > 19 else STRPTIME(txt, FMTDT24HMS)
+
+def td_from_str(txt):
+    """Get a timedelta, given a string"""
+    slen = len(txt)
+    if (slen in {14, 15} and txt[-7] == '.') or slen in {7, 8}:
+        return _td_from_hms(txt, slen)
+    daystr, hms = txt.split(',')
+    return _td_from_hms(hms[1:], len(hms)-1) + \
+           timedelta(days=int(daystr.split(maxsplit=1)[0]))
+
+def _td_from_hms(txt, slen):
+    """Get a timedelta, given 8:00:00 or 12:00:01.100001"""
+    if slen in {14, 15} and txt[-7] == '.':
+        dto = STRPTIME(txt, "%H:%M:%S.%f")
+        return timedelta(hours=dto.hour,
+                         minutes=dto.minute,
+                         seconds=dto.second,
+                         microseconds=dto.microsecond)
+    assert slen in {7, 8}
+    dto = STRPTIME(txt, "%H:%M:%S")
+    return timedelta(hours=dto.hour, minutes=dto.minute, seconds=dto.second)
+
+# ---------------------------------------------------------------------
 def timedelta_to_hms(tdelta):
     """Convert a timedelta to hours and minutes"""
     total_secs = tdelta.total_seconds()
@@ -47,6 +77,7 @@ def _secsneg_to_str(total_secs):
     return f'-{cur_hours:02}:{cur_minutes:02}'
 
 
+# ---------------------------------------------------------------------
 class RoundTime:
     """Round a datetime object up or down to `round_to_min`"""
     # pylint: disable=line-too-long
